update correct row when invalidating student rows
When Gradezilla students are sorted, grading will be broken. As
an example, assume two students (A and B) have been sorted so
that their order is reversed. When student A (now in the second
row) is graded, the assignment group and total grades for
student B (in the first row - student A's original row) will be
updated.

This change might not be immediately apparent, as the update will
only be visible if the grades for the other student had not yet
been updated and displayed.

fixes CNVS-37086

test plan:
1. visit Gradezilla
2. change a submission grade
  a. ensure the student's assignment group grades update correctly
  b. ensure the student's total grade updates correctly
3. sort students so that the given student row has moved
4. change the student's submission grade
  a. ensure the student's assignment group grades update correctly
  b. ensure the student's total grade updates correctly

Change-Id: Ie794c1d17283554348ef9f7110350886e6a54e82
Reviewed-on: https://gerrit.instructure.com/112957
Tested-by: Jenkins
Reviewed-by: Shahbaz Javeed <sjaveed@instructure.com>
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>