Fix a11y navigation modal focus issues
The ticket describes a problem when closing the move dialog from the 'x'
close button, but the same issue exists when closing via the Cancel
button, or selecting Disable or Enable. This commit addresses all the
cases.

Previously, when moving an item from the dialog, focus was put on the
moved item. This commit changes that to be back on the button that
triggered popup menu.  This is where focus is now put for all cases of
the move dialog closing.

fixes CNVS-37153

Test plan:
  - create a course and navigate to it's Settings page, Navigation tab
  - using the settings menu for an item in the list,
    - choose Move. the dialog opens and focus is on the 'x' close button
    - type ENTER to press the button
    > expect focus to return to the cog menu button
    - chose Move
    - TAB to the Cancel button, press ENTER to click it
    > expect focus to return to the cog menu button
    - chose Move
    - Select one of the radio buttons and click the Move button
    > expect the item to move and focus to return to its cog menu button
    - choose Disable
    > expect the item to move to the disabled list and focus to be on
    > its cog menu button
    - on the newly disabled item, choose Enable
    > expect the item to move to the enabled list and focus to be on its
    > cog menu button

Change-Id: Id24e9fd699729cc4b11fc00b60bad111c0194bb6
Reviewed-on: https://gerrit.instructure.com/115227
Tested-by: Jenkins
Reviewed-by: Felix Milea-Ciobanu <fmileaciobanu@instructure.com>
QA-Review: David Tan <dtan@instructure.com>
Product-Review: Pierce Arner <pierce@instructure.com>