Don't show "new feedback" if already seen.
This changes things so that submissions are considered "read" if
the student clicks on an attempt or submission.  Then in the
planner API, we only mark those items that are *not* read as
having "recent" feedback.  This has the effect of no longer
showing the "New Feedback" pill if the student has seen the
assignment/quiz since the last time it was graded (if an
assignment gets graded again, the submission already gets
marked unread, so the "new feedback" pill will return").

Fixes FALCOR-536.

Test Plan:
* Ensure you have https://gerrit.instructure.com/#/c/123039/
* Create a quiz and assignment.
* Have a student take it.
* At this point, in planner, "New Feedback" should not appear.
* Now have it be graded, and make sure to leave some comments.
* Now the pill should appear.
* As the student, go back to the quiz/assignment pages and then
  click on the most recent attempt or submission.
  This should also work from the "grades" tab, and then clicking
  on the assignment/quiz from in there.
* The "New Feedback" and "New Grades" pill should now be gone in
  planner.

Change-Id: I043e64517b8a0873576183800ff67d09161139b7
Reviewed-on: https://gerrit.instructure.com/122740
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: David Mirabile <dmirabile-c@instructure.com>
Product-Review: Christi Wruck