use "enter grades as" setting for display only
closes GRADE-47

QA Notes:

The behavior under test is how grades are formatted in grid cells
according to the "Enter Grades As" setting on each assignment.
This only applies to assignments with grading types which allow
that setting – Points, Percentage, and Letter Grade.

Each verification step will be specific to a combination of the
assignment's "Enter Grades As" setting and the grade applied to
the submission used during verification.

For the Letter Grade assignments, the grading schemes will be
relevant, used to ensure the appropriate scheme key (e.g. A, B, C)
is displayed for the given grade/score. This will be tested only
when the "Enter Grades As" setting is set to "Grading Scheme" for
those assignments.

Note: Cells in "Display mode" will show the "final" version of
the grade. This only matters if late policies are in use on the
course.

test plan:
 A. Setup
    1. Select or create a course
    2. Add a custom grading scheme to the course
       * ensure this is distinguishable from the
         default grading scheme on the course/account
    3. Select or create assignments with grading types:
       a. Points
       b. Percentage
       c. Letter Grade (with default grading scheme)
       d. Letter Grade (with the custom grading scheme)
       e. Complete/Incomplete
 B. Verification
    1. Visit New Gradebook
    2. For each assignment and "Enter Grades As" option, verify:
       a. the grade is displayed in the expected format
       b. the grade is accurately converted between formats
          * e.g. 10 / 10 is also 100%, which is also "A"
    3. Smoke-test grade entry for each assignment type

Change-Id: Ia366eaad7c595549a6e8602775e60a68515ff7cf
Reviewed-on: https://gerrit.instructure.com/129940
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Shahbaz Javeed <sjaveed@instructure.com>
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Matt Goodwin <mattg@instructure.com>