Make it work with the version of React Canvas uses
The version of react that canvas uses (“0.14.9”) does not have
React.PureComponent. so this will throw an error.

The reason why this was not caught was because the initial commit
That added this to canvas had changes that it did not commit to
yarn.lock and so in the review, no one noticed that it was making
it add a new copy of react to yarn.lock to use just for this. And that
New copy of react (15.x) did have `PureComponent`

Test plan:
* when this is used in canvas, it should work

Change-Id: Ie56e7c5bc47923b4144be61396ec73ef20563d94
Reviewed-on: https://gerrit.instructure.com/129323
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>