handle column resizing with GradebookGrid event
This moves the callback logic for column resizing to three
different layers of Gradebook.

In GridSupport, the columns with new widths are isolated and sent
to the subscriber. If all columns have unchanged widths, no
subscribers are notified.

In GradebookGrid, the column definitions are updated to correct
for a bug that reverts column widths when the columns are
re-rendered.  Additional, Gradebook-specific logic will be
defined here in subsequent commits for related concerns.

In Gradebook, updated column widths are saved via the API.

fixes GRADE-399

test plan:
 1. Visit New Gradebook
 2. Resize a column
 3. Select View > Arrange By > *
    * Select any option
 4. Verify the resized column retains its new width
 5. Smoke test column resizing

Change-Id: I60f4a1591fcfeae3ea7becaec5f784575f15436b
Reviewed-on: https://gerrit.instructure.com/131861
Tested-by: Jenkins
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>