Make course home page go to modules if wiki home page unset.
When a wiki front page is set to the front page of a course, make
the wiki page go to the default for the context (by clearing out
to nil).  In the case of a course home page, this will be the
wiki.

Fixes COMMS-486

Test Plan:
* Have a course.
* Have a page for it, make it the front page.  Set this to further
  be the home page for the course.
* Now unset the page as the "front page" (go to pages->view all
  pages and use the cog next to the page)
* Go to the course home page.
* It should be modules!

Change-Id: Ie2b5e5d44d3d2f6077b193289b385cb58811e261
Reviewed-on: https://gerrit.instructure.com/132314
Tested-by: Jenkins
Reviewed-by: Aaron Kc Hsu <ahsu@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Venk Natarajan <vnatarajan@instructure.com>