Remove "Prefer Office 365 file viewer" setting
Fixes RD-4501

This option currently does not work.  We do intend to eventually get this
working, but until then, we'll remove the option.  This is commented out
rather than removed completely because we intend to re-introduce it, and
we don't want anyone removing code critical to this in the meantime.

If there is anyone with this option currently enabled we might have to do
data cleanup for them.

Test plan:
- Load /accounts/{accountId}/settings
- Confirm that the "Prefer Office 365 File Viewer" setting is no longer
present

Change-Id: I61d48854b54acacb79853f75485d9920f40705f0
Reviewed-on: https://gerrit.instructure.com/134110
Tested-by: Jenkins
Reviewed-by: brian kirkby <bkirkby@instructure.com>
Product-Review: brian kirkby <bkirkby@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>