fix bug where rce image plugin removes links
calling insertContent on a tinymce editor will replace the link
surrounding an img if only an img is selected. if the link plugin is
used when the selected node is an image, the update function will
update the attributes of the existing img element instead of trying
to replace it.

closes CORE-640

test plan:
- make sure rcs feature flag is enabled
- insert an image in a rce
- make the image a link
- select the image
- click the image button in the toolbar
- change something (i.e dimensions or alt text)
- click update
- the image should still be a link

Change-Id: Ie0ab797d013f88261bbca280f15493bdef1ec94c
Reviewed-on: https://gerrit.instructure.com/134130
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Tucker McKnight <tmcknight@instructure.com>
Product-Review: Brent Burgoyne <bburgoyne@instructure.com>