gradebook group submission comments->radio buttons
Convert the gradebook submission dialog group comments from a checkbox
to the radio buttons.

This also fixes a bug where the dialog's content was left in the DOM
each time the dialog was opened. This caused multiple elements with the
same id to exist on the page, which caused the radio buttons to not
work correctly the second time the dialog was opened.

closes ADMIN-661
part of ADMIN-578

test plan:
* as a teacher, create an individually graded group assignment and
  go to the gradebook.
* click on the comment corner of a gradebook cell for that assignment
* the dialog should render radio buttons for sending the submission
  comment to the student or the group
* both values of the radio button should work as expected
* open the dialog more than once and verify the radio buttons are still
  clickable

* create a group graded group assignment
* click on the comment corner of a gradebook cell for that assignment
* the dialog should render a message saying all comments go to the group
* a comment posted should go to the whole group

Change-Id: I483e6ecccd2638b230fb4464df0790cb5d769ae8
Reviewed-on: https://gerrit.instructure.com/137968
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Christi Wruck
Tested-by: Jenkins
QA-Review: Mysti Sadler <mysti@instructure.com>