[npm dependencies] update all our npm dependencies
This is just the regular “update all the things” commit.
For things in package.json that we just want the latest version of,
I changed them to a simple major with a carat. (Eg: “8^” or “4^”),
that way you can see easier which things are specifically locked down
To a specific version (or range of versions) for a reason.

Test plan:
* automated specs & build tasks should work

Change-Id: Id7a3f853bd0f7537cc0f08d7dba8d6726b836c44
Reviewed-on: https://gerrit.instructure.com/141653
Tested-by: Jenkins
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
QA-Review: Brent Burgoyne <bburgoyne@instructure.com>
Product-Review: Brent Burgoyne <bburgoyne@instructure.com>