Add search and filter to discussions index v2
Note: You may see performance is not so great when filtering lots of
discussions. The bottleneck is the react rendering lots of components,
not the actual filtering we are doing. I'm hoping the performance will
be better when the javascript is build for production, but it's
something we should test in the performance checklist.

Fixes COMMS-944

Test Plan:
  * Go to a course, and create some discussions with different users.
  * Notice that you can search for discussions based on their name
  * Notice you can filter discussions via read/unread
  * Notice you can combine filtering and searching together, and
    everything will work as expected.
  * Test with keyboard only naviagation and make sure everything is
    accessable
  * Test with a screen reader and notice that it tells you how many
    results were found, and that if you are typing quickly it only
    tells you the results once you stop typing

Change-Id: I88c91dbdc6ebc3b5206caf56af8cea4111da5c06
Reviewed-on: https://gerrit.instructure.com/143012
Reviewed-by: Aaron Kc Hsu <ahsu@instructure.com>
Reviewed-by: Venk Natarajan <vnatarajan@instructure.com>
Tested-by: Jenkins
Product-Review: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>