remove webpackHooks, I don’t think anyone uses it
This was a thing so you could get Mac notifications of your webpack 
build progress. I don’t think anyone actually uses it though.

I’ll add a bunch of people to the review on this and if you or someone
You know uses it still, let me know and we can keep it in if it is 
helpful.

Change-Id: Iebf7fb16ebd532e63cbb31b325f7cd8e1b711e02
Reviewed-on: https://gerrit.instructure.com/143257
Tested-by: Jenkins
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>