Change handling of loading for pagination
This makes it so rather than replace the whole pagination
component with a loading message, we put a loading spinner
in the place of the page that is to become the current page.

This keeps the keyboard focus on the component in the proper
location rather than starting from the beginning of the component
on each re-render.

closes CORE-1108

Test Plan:
  - Go to the new account user/course search
  - Make sure on one of those pages you have enough
    users/courses to have several pages (75 or so should be
    enough to force a decent number of pages)
  - Use a keyboard to adjust the page
  - A spinner should show on the item you plan to go to
  - Focus should be on the element after the spinner changes out to
    a number.

Change-Id: I3d4be7d440260bc1941f3084781535c029874b27
Reviewed-on: https://gerrit.instructure.com/143923
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>