fix: Add People modal table needs row headers
Closes: CORE-1120

Test plan:
When a screen reader user is adding people to a course through the 
Add People modal, the table on the second page has no row headers. This 
means that the user has to move focus back to the Name column to know 
what any of the data means if they navigate up/down through the results.

We should use the Name column in this case.

Change-Id: I3c2a8e95787ba0ffcb0483b3d7313a634b6a4850
Reviewed-on: https://gerrit.instructure.com/144307
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>