BBB enable Arc integration
Summary:
Extension for BigBluebutton recordings ready post-processing
that enables the integration with Arc.

Description:
The only update is to the big_blue_button model for adding a new meta
parameter that includes the user email in RFC2822 format. This parameter
is required in order to match the user accounts between Canvas and Arc.

Test Plan:
As everything happens in the BigBlueButton recording server, the only
thing to test here is that the parameter is included and correctly sent.

Development ID:
blindsidenetworks#22

Internally ID:
Closes COMMS-604

Change-Id: I330746525013b846cbd79ee1c32626e6cc40408d
Reviewed-on: https://gerrit.instructure.com/146018
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Tested-by: Jenkins
QA-Review: Kyle Rosenbaum <krosenbaum@instructure.com>
Product-Review: Matt Goodwin <mattg@instructure.com>