Correctly scroll to most recent new activity
when the planner is in tablet mode, the new activity indicators are next
to the item, not the group. the previous code didn't take that into
account when it looked for it by sorting on animatableIndex.
Unfortunately, there can be duplicate PlannerItem animatableIndices.
Deal with that by having PlannerApp -> Day -> Grouping -> PlannerItem
assign the index using this.props.animatableIndex * 100 + myIndex. A
nice side effect of this is that you can look in the animatable-registry
and see the logical groupings.

fixes ADMIN-911

test plan:
It was a fluke that my local date generated PlannerItems with new
activity with duplicate animatableIndex values. That's the only way
I stumbled on this issue.
  - Have items due with new activity in due the past,
    a couple in the same day but different courses.
  - load the planner, and click the New Activity button
  > expect it to scroll to the most recent in the past item with new
  > activity and focus on the checkbox
  - check the checkbox so the new activity item is completed
  - refresh the browser
  - click New Activity
  > expect it to scroll to the item and focus on the "show 1 completed
  > item" toggle.
Also
  - ensure that the item with new activity receives focus

Change-Id: I2cf5fbdcb5cf43647c5dce121149c03ae72a94bf
Reviewed-on: https://gerrit.instructure.com/149728
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>