change assignment due date caching from a singleton to a strand
When invoking the DueDateCacher with update_grades on, switch to using
a strand based on the course instead of a singleton based on the
assignment. This should avoid score table contention when lots of
assignments in a single course are edited close in time to each other.

closes GRADE-1194

test plan:
 - Have a course with some students and a bunch of assignments
 - Edit the assignments quickly back to back
 - Ensure the jobs queued as a strand
 - (Testing this may be easier if delayed jobs isn't running and you
   edit inst-jobs to always background even if not in production.)

Change-Id: I4b5600f52f2f09b6faf994e29e6056e1c6eaeb60
Reviewed-on: https://gerrit.instructure.com/151661
Tested-by: Jenkins
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Jeremy Neander <jneander@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>