canvas-rce: 4.0.0
We have made a bunch of commits to canvas-rce since we last did a
Release to npm so this is so we can publish the latest code to npm so
that I can consume it (and all the instUI 5 changes) in 
tinymce-a11y-checker. Since a lot have changed I figured I’d better make
it a major version bump.

Change-Id: I549e85a207ba0299c63262cac9710a95b52c71c8
Reviewed-on: https://gerrit.instructure.com/154403
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>